# framework/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rover steer service",
        "description": "Rover steer service",
        "contact": {
            "email": "omer.faruk.kesmez@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost/api/v2",
            "description": "Rover steer service"
        },
        {
            "url": "https://localhost/api/v1",
            "description": "Rover steer service"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "TOKEN"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "info@okesmez.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                }
            }
        },
        "/plateau": {
            "get": {
                "tags": [
                    "Plateau"
                ],
                "summary": "Get Plateau List.",
                "operationId": "getPlateauList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plateau"
                ],
                "summary": "Create Plateau.",
                "operationId": "createPlateau",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/plateau/{id}": {
            "get": {
                "tags": [
                    "Plateau"
                ],
                "summary": "Show Plateau.",
                "operationId": "showPlateau",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/rover": {
            "get": {
                "tags": [
                    "Rover"
                ],
                "summary": "Get rover List.",
                "operationId": "getroverList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rover"
                ],
                "summary": "Create Rover.",
                "operationId": "createRover",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "plateau_id": {
                                        "type": "integer"
                                    },
                                    "direction": {
                                        "type": "string"
                                    },
                                    "x": {
                                        "type": "integer"
                                    },
                                    "y": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "string",
                                    "plateau_id": "integer",
                                    "direction": "<N,S,W,E>",
                                    "x": "integer",
                                    "y": "integer"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/rover/{id}": {
            "get": {
                "tags": [
                    "Rover"
                ],
                "summary": "Show Rover.",
                "operationId": "showRover",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/state": {
            "post": {
                "tags": [
                    "State"
                ],
                "summary": "Create State.",
                "operationId": "createState",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number_of_progress": {
                                        "type": "integer"
                                    },
                                    "rover_id": {
                                        "type": "integer"
                                    },
                                    "rover_rotation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "number_of_progress": "Düzlemde ilerleme miktarı (integer)",
                                    "rover_id": "integer",
                                    "rover_rotation": "<R,L>"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/rover/state/{rover_id}": {
            "get": {
                "tags": [
                    "State"
                ],
                "summary": "Show Rover State.",
                "operationId": "showRoverState",
                "parameters": [
                    {
                        "name": "rover_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Not successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>) For Example : Bearer adasdasdsaplıfmapsfaapfmd",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}